requirements.txt
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-background-tasks==1.2.0
django-compat==1.0.15
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
pep8==1.7.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python-memcached==1.59
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
Unidecode==1.0.23
urllib3==1.24.2
wrapt==1.11.1
